LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/reporting-to-document-reporting-endpoint.https.window-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/text-plain.txt?16dde056-b87c-48bc-aa9d-f8c11268f6b6 because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/text-plain.txt?16dde056-b87c-48bc-aa9d-f8c11268f6b6 due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html?168a51fb-d093-460b-b0fd-e4b35d616bf8 because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/common/blank.html?ebafb0be-4259-410b-a6bb-b7926296e51e' in a frame because of Cross-Origin-Embedder-Policy.

FAIL subresource CORP assert_true: A corp report with blockedURL https://127.0.0.1:9443/common/text-plain.txt and url https://localhost:9443/html/cross-origin-embedder-policy/resources/reporting-empty-frame.html?pipe=header(cross-origin-embedder-policy,require-corp;report-to=%22endpoint%22)|header(cross-origin-embedder-policy-report-only,require-corp;report-to=%22report-only-endpoint%22)|header(reporting-endpoints,%20endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=05309f37-6415-4981-b413-c86d603b2680%22\,%20report-only-endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=ffdbed22-f513-4199-b829-43d9a6cbef15%22) is not found. expected true got false
FAIL navigation CORP on cross origin assert_true: A corp report with blockedURL https://127.0.0.1:9443/common/blank.html and url https://localhost:9443/html/cross-origin-embedder-policy/resources/reporting-empty-frame.html?pipe=header(cross-origin-embedder-policy,require-corp;report-to=%22endpoint%22)|header(cross-origin-embedder-policy-report-only,require-corp;report-to=%22report-only-endpoint%22)|header(reporting-endpoints,%20endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=05309f37-6415-4981-b413-c86d603b2680%22\,%20report-only-endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=ffdbed22-f513-4199-b829-43d9a6cbef15%22) is not found. expected true got false
FAIL navigation CORP on same origin assert_true: A navigation report with blockedURL https://localhost:9443/common/blank.html and url https://localhost:9443/html/cross-origin-embedder-policy/resources/reporting-empty-frame.html?pipe=header(cross-origin-embedder-policy,require-corp;report-to=%22endpoint%22)|header(cross-origin-embedder-policy-report-only,require-corp;report-to=%22report-only-endpoint%22)|header(reporting-endpoints,%20endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=05309f37-6415-4981-b413-c86d603b2680%22\,%20report-only-endpoint=%22/html/cross-origin-embedder-policy/resources/report.py?key=ffdbed22-f513-4199-b829-43d9a6cbef15%22) is not found. expected true got false
